LayoutTests/imported/w3c/web-platform-tests/navigation-timing/content-encoding.https_pipe=gzip-expected.txt

Description

This test validates that when a html is compressed with gzip, navigation timing reports contentEncoding as gzip


FAIL contentEncoding should be gzip. assert_equals: Expected contentEncoding to be gzip. expected (string) "gzip" but got (undefined) undefined
